difference of gaussians
figured out how to do difference of gaussians.  It looks like applying
this for edge detection, or for our application in-map competition
involves generating a diff of gauss. kernel and convolving it with the
feature map as usual.  In koch et al, the resulting map is then
rectified to remove all negative values and then iterated over a
sequence of the same convolution on a single map.